Swagger added. Works now
