+++ src/main/java/educationalproject/programmingstuff/ProgrammingStuffApplication.java
@@ -1,17 +1,32 @@
 package educationalproject.programmingstuff;
 
-import educationalproject.programmingstuff.service.MainService;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.ConfigurableApplicationContext;
+import org.springframework.context.annotation.Bean;
+import springfox.documentation.builders.PathSelectors;
+import springfox.documentation.builders.RequestHandlerSelectors;
+import springfox.documentation.spi.DocumentationType;
+import springfox.documentation.spring.web.plugins.Docket;
+import springfox.documentation.swagger2.annotations.EnableSwagger2;
+
+//todo swagger
+//todo design and make relationships between entities
 
 @SpringBootApplication
+@EnableSwagger2
 public class ProgrammingStuffApplication {
 
     public static void main(String[] args) {
-        ConfigurableApplicationContext applicationContext = SpringApplication.run(ProgrammingStuffApplication.class, args);
-        MainService mainService=(MainService) applicationContext.getBean("MainService");
-        mainService.prepareTestData();
+        SpringApplication.run(ProgrammingStuffApplication.class, args);
+    }
+
+    @Bean
+    public Docket api() {
+        return new Docket(DocumentationType.SWAGGER_2)
+                .select()
+                .apis(RequestHandlerSelectors.any())
+                .paths(PathSelectors.any())
+                .build();
     }
 
 }
