+++ src/main/java/educationalproject/programmingstuff/model/Order1.java
@@ -1,4 +1,4 @@
-package educationalproject.programmingstuff.service.model;
+package educationalproject.programmingstuff.model;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -9,21 +9,18 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import java.math.BigDecimal;
 
 @Entity
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-
-public class Item {
+public class Order1 {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private String title;
-    private String description;
-    private BigDecimal price;
+    private Long userId;
+
 }
