+++ src/main/java/educationalproject/programmingstuff/repository/ItemRepository.java
@@ -1,6 +1,6 @@
 package educationalproject.programmingstuff.repository;
 
-import educationalproject.programmingstuff.service.model.Item;
+import educationalproject.programmingstuff.model.Item;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
