+++ src/main/java/educationalproject/programmingstuff/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package educationalproject.programmingstuff.repository;
 
-import educationalproject.programmingstuff.service.model.User;
+import educationalproject.programmingstuff.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
