+++ src/main/resources/application.yml
@@ -1,0 +1,4 @@
+spring:
+  mvc:
+    pathmatch:
+      matching-strategy: ant_path_matcher
